+++ lib/commands/index.js
@@ -1,0 +1,27 @@
+const { start } = require('./start')
+const { select } = require('./select')
+const { status } = require('./status')
+const { help } = require('./help')
+
+const { listen } = require('./listen')
+const { send } = require('./send')
+const { filter } = require('./filter')
+const { equip } = require('./equip')
+const { quit } = require('./quit')
+const { consume, inventory, tossItem, useItem } = require('./inventory')
+
+module.exports = {
+  start: start,
+  listen: listen,
+  select: select,
+  send: send,
+  status: status,
+  filter: filter,
+  inventory: inventory,
+  equip: equip,
+  consume: consume,
+  use_item: useItem,
+  toss: tossItem,
+  help: help,
+  quit: quit
+}
