+++ lib/commands/status.js
@@ -1,0 +1,6 @@
+function status (ctx) {
+  let status = ctx.db.getStatus()
+  ctx.reply(status)
+}
+
+module.exports = status
